requirements.txt
pandas
pymongo
numpy
matplotlib
seaborn
statsmodels
scikit-learn
pyyaml  
scipy
imblearn
xgboost
catboost
certifi
evidently
dill
-e .